refactor(jolt-crypto): adapt to main's lint rules and module layout
- Rename src/arkworks/ → src/ec/ to match local refactor/crates layout
- Replace #[allow] with #[expect] per main's strict lint config
- Add #[expect(clippy::expect_used/unwrap_used)] where needed
- Remove serialized_len from AppendToTranscript (not in main's trait)
- Delete README.md and REVIEW.md (not shipped to main)
