--- crates/jolt-crypto/REVIEW.md
@@ -1,94 +1,0 @@
-# jolt-crypto Review
-
-**Crate:** jolt-crypto (Level 2)
-**LOC:** 4,969 (was ~5,070 — reduced via G1/G2 macro dedup)
-**Baseline:** 0 clippy warnings, 128 tests passing
-**Rating:** 8.5/10
-
-## Overview
-
-Elliptic curve and commitment abstractions for Jolt. Provides the `JoltGroup`,
-`PairingGroup`, `JoltCommitment`, and `HomomorphicCommitment` trait hierarchy
-with a BN254 backend. Includes GLV scalar multiplication, batch affine addition,
-Pedersen commitments, and Dory interop bridge. Zero arkworks leakage in the
-public API — all arkworks types are behind `#[repr(transparent)]` newtypes.
-
-**Verdict:** Clean trait design with strong encapsulation. The `#[repr(transparent)]`
-newtypes effectively hide arkworks internals. Performance-critical paths (GLV, batch
-addition, MSM) are well-optimized. The G1/G2 macro dedup eliminated 300+ lines of
-identical boilerplate. Well-tested with 128 tests, 14 benchmarks, and 3 fuzz targets.
-
----
-
-## Findings
-
-### [CQ-1.1] G1/G2 ~300 LOC identical boilerplate
-**File:** `src/arkworks/bn254/g1.rs`, `g2.rs`
-**Severity:** MEDIUM
-**Finding:** Both files had 157 LOC of nearly identical operator impls, serde, JoltGroup impl.
-**Status:** RESOLVED — Created `impl_jolt_group_wrapper!` macro in `mod.rs`. Both files reduced to 8 LOC.
-
-### [CQ-2.1] Missing compile-time size assertions for repr(transparent)
-**File:** `src/arkworks/bn254/g1.rs`, `g2.rs`
-**Severity:** HIGH
-**Finding:** Unsafe pointer casts between wrapper and inner types rely on `#[repr(transparent)]` layout guarantee but had no compile-time verification.
-**Status:** RESOLVED — Added `const _: () = assert!(size_of::<Wrapper>() == size_of::<Inner>());` inside the macro.
-
-### [CQ-2.2] No safe into_inner() accessor
-**File:** `src/arkworks/bn254/g1.rs`, `g2.rs`
-**Severity:** LOW
-**Finding:** Downstream crates (jolt-dory) use unsafe transmute to extract the inner arkworks type. A safe accessor would reduce unsafe surface area.
-**Status:** RESOLVED — Added `pub fn into_inner(self)` to the macro. Downstream migration (jolt-dory scheme.rs ~15 transmutes) deferred.
-
-### [CQ-3.1] Clippy warnings in dory_interop
-**File:** `src/dory_interop.rs`
-**Severity:** LOW
-**Finding:** 3 redundant closures, 1 needless borrow.
-**Status:** RESOLVED — Fixed all 4 warnings.
-
-### [CQ-4.1] Missing #[must_use] on pure trait methods
-**File:** `src/groups.rs`, `src/commitment.rs`
-**Severity:** LOW
-**Finding:** Pure methods returning values without side effects should be `#[must_use]`.
-**Status:** RESOLVED — Added to JoltGroup, JoltCommitment, HomomorphicCommitment, PairingGroup.
-
-### [CD-1.1] Internal modules exposed in public API
-**File:** `src/arkworks/bn254/mod.rs`
-**Severity:** LOW
-**Finding:** `glv` and `batch_addition` modules are `pub` but are implementation details.
-**Status:** RESOLVED — Added `#[doc(hidden)]`.
-
-### [CD-2.1] batch_addition precondition undocumented
-**File:** `src/arkworks/bn254/batch_addition.rs`
-**Severity:** LOW
-**Finding:** Entry point has subtle preconditions (in-bounds indices, no equal/inverse pairs).
-**Status:** RESOLVED — Added doc comment.
-
-### [CD-3.1] bincode v1 API usage
-**File:** `tests/serialization.rs`, `benches/crypto.rs`, `fuzz/fuzz_targets/*.rs`
-**Severity:** MEDIUM
-**Finding:** Workspace uses bincode v2 but these files used v1 API.
-**Status:** RESOLVED — Migrated all files + fuzz Cargo.toml to bincode v2.
-
-### [CD-4.1] Cargo.toml metadata incomplete
-**File:** `Cargo.toml`
-**Severity:** LOW
-**Finding:** Missing repository, keywords, categories; license was MIT-only.
-**Status:** RESOLVED — Updated to dual MIT OR Apache-2.0, added all metadata.
-
-### [NIT-1.1] append_to_transcript Vec allocation
-**File:** `src/arkworks/bn254/g1.rs` (now in macro)
-**Severity:** LOW
-**Finding:** `ark_serialize::CanonicalSerialize` allocates a fresh Vec on each call.
-**Status:** WONTFIX — Acceptable for now; not on hot path.
-
----
-
-## Summary
-
-| Category | Pass | Resolved | Wontfix | Total |
-|----------|------|----------|---------|-------|
-| CQ       | 0    | 4        | 0       | 4     |
-| CD       | 0    | 4        | 0       | 4     |
-| NIT      | 0    | 0        | 1       | 1     |
-| **Total**| **0**| **8**    | **1**   | **9** |
