+++ crates/jolt-crypto/benches/crypto.rs
@@ -1,9 +1,9 @@
-#![allow(unused_results)]
+#![expect(unused_results, clippy::unwrap_used)]
 
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 
 use jolt_crypto::{
-    Bn254, Bn254G1, Bn254G2, JoltCommitment, JoltGroup, PairingGroup, Pedersen, PedersenSetup,
+    Bn254, Bn254G1, Bn254G2, JoltGroup, PairingGroup, Pedersen, PedersenSetup, VectorCommitment,
 };
 use jolt_field::{Field, Fr};
 use rand_chacha::ChaCha20Rng;
