--- crates/jolt-crypto/src/arkworks/mod.rs
@@ -1,5 +1,0 @@
-//! Arkworks backend implementations.
-//!
-//! This module contains group implementations using the arkworks library.
-
-pub mod bn254;
