+++ crates/jolt-crypto/src/ec/bn254/glv/power_of_2_decompositions.rs
