+++ crates/jolt-crypto/src/ec/mod.rs
@@ -1,0 +1,13 @@
+//! Elliptic curve primitives: groups, pairings, and Pedersen commitments.
+
+mod group;
+pub use group::JoltGroup;
+
+mod pairing;
+pub use pairing::PairingGroup;
+
+mod pedersen;
+pub use pedersen::{Pedersen, PedersenSetup};
+
+#[cfg(feature = "bn254")]
+pub mod bn254;
