+++ crates/jolt-crypto/tests/coverage.rs
@@ -3,7 +3,7 @@
 //! Covers gaps in G2, GT, GLV (2D/4D), Dory vector ops, fixed-base MSM,
 //! HomomorphicCommitment, and Debug/From conversions.
 
-use jolt_crypto::arkworks::bn254::glv;
+use jolt_crypto::ec::bn254::glv;
 use jolt_crypto::{
     Bn254, Bn254G1, Bn254G2, Bn254GT, HomomorphicCommitment, JoltGroup, PairingGroup,
 };
@@ -32,7 +32,7 @@ fn g2_commutativity() {
 }
 
 #[test]
-#[allow(clippy::op_ref)]
+#[expect(clippy::op_ref)]
 fn g2_add_ref() {
     let g = Bn254::g2_generator();
     let a = g.scalar_mul(&Fr::from_u64(3));
@@ -42,7 +42,7 @@ fn g2_add_ref() {
 }
 
 #[test]
-#[allow(clippy::op_ref)]
+#[expect(clippy::op_ref)]
 fn g2_sub_ref() {
     let g = Bn254::g2_generator();
     let a = g.scalar_mul(&Fr::from_u64(7));
@@ -147,7 +147,7 @@ fn gt_sub_assign() {
 }
 
 #[test]
-#[allow(clippy::op_ref)]
+#[expect(clippy::op_ref)]
 fn gt_add_ref() {
     let e = gt_element();
     let e2 = e.scalar_mul(&Fr::from_u64(2));
@@ -156,7 +156,7 @@ fn gt_add_ref() {
 }
 
 #[test]
-#[allow(clippy::op_ref)]
+#[expect(clippy::op_ref)]
 fn gt_sub_ref() {
     let e = gt_element();
     let e2 = e.scalar_mul(&Fr::from_u64(2));
