+++ crates/jolt-crypto/tests/serialization.rs
@@ -1,3 +1,4 @@
+#![expect(clippy::expect_used)]
 //! Serialization round-trip tests for all BN254 types.
 
 use jolt_crypto::{Bn254, Bn254G1, Bn254G2, Bn254GT, JoltGroup, PairingGroup, PedersenSetup};
